backend/requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
azure-cognitiveservices-speech==1.16.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.5
virtualenv==20.4.3
Werkzeug==1.0.1
